requirements.txt
streamlit
ultralytics
opencv-python
numpy
pandas
scikit-learn
scipy
matplotlib
seaborn
pillow
requests
tqdm